refactor most checkers
