+++ argopy/utils/__init__.py
@@ -1,7 +1,20 @@
 from .monitored_threadpool import MyThreadPoolExecutor as MonitoredThreadPoolExecutor
-
+from .checkers import (
+    is_box, is_indexbox,
+    is_list_of_strings, is_list_of_dicts, is_list_of_datasets, is_list_equal,
+    is_wmo, check_wmo,
+    is_cyc, check_cyc,
+    check_index_cols,
+)
 
 __all__ = (
     # Classes:
     "MonitoredThreadPoolExecutor",
+
+    # Checkers:
+    "is_box", "is_indexbox",
+    "is_list_of_strings", "is_list_of_dicts", "is_list_of_datasets", "is_list_equal",
+    "is_wmo", "check_wmo",
+    "is_cyc", "check_cyc",
+    "check_index_cols",
 )
